Added support for browser journal integration to PageFrame Updated NewsWidget sample to use journal integration

+++ src/Client/Core/UserInterface/Navigation/PageJournal.cs
@@ -10,6 +10,9 @@
 
 using System;
 using System.Collections.Generic;
+using System.Windows;
+using System.Windows.Browser;
+using System.Windows.Interop;
 
 namespace SilverlightFX.UserInterface.Navigation {
 
@@ -18,6 +21,10 @@ namespace SilverlightFX.UserInterface.Navigation {
         private List<Uri> _journalEntries;
         private int _current;
 
+        private bool _addingJournalEntry;
+        private bool _integrated;
+        private Action<Uri> _navigateAction;
+
         public PageJournal() {
             _journalEntries = new List<Uri>();
             _current = -1;
@@ -25,26 +32,60 @@ namespace SilverlightFX.UserInterface.Navigation {
 
         public bool CanGoBack {
             get {
+                if (_integrated) {
+                    return true;
+                }
                 return _current > 0;
             }
         }
 
         public bool CanGoForward {
             get {
+                if (_integrated) {
+                    return true;
+                }
                 return _current < _journalEntries.Count - 1;
             }
         }
 
-        public void AddEntry(Uri uri) {
-            if (_current != _journalEntries.Count - 1) {
-                _journalEntries.RemoveRange(_current, _journalEntries.Count - _current);
+        public bool CanIntegrateWithBrowser {
+            get {
+                return HtmlPage.IsEnabled;
             }
+        }
 
-            _journalEntries.Add(uri);
-            _current = _journalEntries.Count - 1;
+        public bool IsIntegratedWithBrowser {
+            get {
+                return _integrated;
+            }
+        }
+
+        public void AddEntry(Uri uri) {
+            if (_integrated) {
+                try {
+                    _addingJournalEntry = true;
+                    Application.Current.Host.NavigationState = uri.ToString();
+                }
+                finally {
+                    _addingJournalEntry = false;
+                }
+            }
+            else {
+                if (_current != _journalEntries.Count - 1) {
+                    _journalEntries.RemoveRange(_current, _journalEntries.Count - _current);
+                }
+
+                _journalEntries.Add(uri);
+                _current = _journalEntries.Count - 1;
+            }
         }
 
         public Uri GoBack() {
+            if (_integrated) {
+                ((ScriptObject)HtmlPage.Window.GetProperty("history")).Invoke("back");
+                return null;
+            }
+
             if (_current > 0) {
                 _current--;
                 return _journalEntries[_current];
@@ -53,11 +94,31 @@ namespace SilverlightFX.UserInterface.Navigation {
         }
 
         public Uri GoForward() {
+            if (_integrated) {
+                ((ScriptObject)HtmlPage.Window.GetProperty("history")).Invoke("forward");
+                return null;
+            }
+
             if (_current < _journalEntries.Count - 1) {
                 _current++;
                 return _journalEntries[_current];
             }
             return null;
+        }
+
+        public void IntegrateWithBrowser(Action<Uri> navigateAction) {
+            _navigateAction = navigateAction;
+
+            Application.Current.Host.NavigationStateChanged += OnHostNavigationStateChanged;
+            _integrated = true;
+        }
+
+        private void OnHostNavigationStateChanged(object sender, NavigationStateChangedEventArgs e) {
+            if (_addingJournalEntry) {
+                return;
+            }
+
+            _navigateAction(new Uri(e.NewNavigationState, UriKind.Relative));
         }
     }
 }
